Update of the action column of the Suptech model from the LOG_SUPTECH.xls file
