--- tools/management/commands/_csv_suptech.py
@@ -1,33 +1,0 @@
-from utils.microsoft_format import CsvFormat
-
-
-class CsvSuptech(CsvFormat):
-    COLS = {"DATE": "date", "QUI": "user", "XELON": "xelon", "ITEM": "item", "TIME": "time", "INFO": "info",
-            "RMQ": "rmq"}
-    COLS_DATE = {'date': "%d/%m/%Y"}
-
-    def __init__(self, file, columns=None):
-        """
-        Initialize ExcelSqualaetp class
-        :param file:
-            excel file to process
-        :param sheet_index:
-            Sheet index to be processed from excel file
-        :param columns:
-            Number of the last column to be processed
-        """
-        super(CsvSuptech, self).__init__(file, columns, dtype=str, usecols=self.COLS.keys())
-        self._columns_rename(self.COLS)
-        self.sheet.replace({"": None}, inplace=True)
-        self._date_converter(self.COLS_DATE)
-
-    def read(self):
-        """
-        Formatting data from the Raspeedi excel file
-        :return:
-            list of dictionnaries that represents the data in the sheet
-        """
-        data = []
-        for line in range(self.nrows):
-            data.append(dict(self.sheet.loc[line].dropna()))
-        return data
